+++ src/ios/frameworks/iOS/Appboy_iOS_SDK.framework/Headers/ABKNavigationBar.h
@@ -1,0 +1,8 @@
+#import <UIKit/UIKit.h>
+
+/*
+ * Braze Public API: ABKNavigationBar
+ */
+@interface ABKNavigationBar : UINavigationBar
+
+@end