+++ src/ios/frameworks/iOS/Appboy_iOS_SDK.framework/Headers/ABKUIUtils.h
@@ -1,0 +1,12 @@
+#import <Foundation/Foundation.h>
+
+@interface ABKUIUtils : NSObject
+
++ (NSString *)getLocalizedString:(NSString *)key inAppboyBundle:(NSBundle *)appboyBundle table:(NSString *)table;
++ (BOOL)objectIsValidAndNotEmpty:(id)object;
++ (UIImage *)maskImage:(UIImage *)image toColor:(UIColor*)color;
++ (Class)getSDWebImageProxyClass;
++ (Class)getModalFeedViewControllerClass;
++ (BOOL)isiPhoneX;
+
+@end